Create README.md
SAP EHS stands for Environment Health and Safety. It is used for dealing with risks related to the environment like pollution, employee health & safety, accidents, etc. SAP EHS is a software that was developed to effectively manage occupational health, industrial hygiene, and the overall safety of employees, products, and the organization. It enables the integration of safety and health tasks in the organization's existing business processes. The software enables organizations to be legally compliant, decrease costs, and lower risks at the same time. It provides a comprehensive solution for the management of the environment, health, and safety of the overall organization.